Small changes to make it work with MSVC.

+++ src/flappy_box/view/paddle_al_audible.cpp
@@ -17,8 +17,8 @@ PaddleAlAudible::PaddleAlAudible( std::shared_ptr< ::flappy_box::model::Paddle >
     alGenSources(1, &source);
     alSourcei(source, AL_BUFFER, buffer);
     alSourcei(source, AL_LOOPING, true);
-    alSourcef(source, AL_GAIN, 0.4);
-    alSourcef(source, AL_PITCH, 0.5);
+    alSourcef(source, AL_GAIN, 0.4f);
+    alSourcef(source, AL_PITCH, 0.5f);
     
     alSourcePlay(source);
 }
@@ -29,13 +29,13 @@ void PaddleAlAudible::auralize( ::view::AlRenderer& renderer )
     {
         // normalized x-coordinate of box
         x_pos = _model->position()[0] / _model->maxPosition()[0];
-        alSource3f(source, AL_POSITION, x_pos, 0., 0.);
+        alSource3f(source, AL_POSITION, static_cast<float>(x_pos), 0.0f, 0.0f);
         
         // pitch = speed
         vec3_type speed = _model->velocity();
         double p = ((speed(0)>0)?speed(0):-speed(0)) / 100. + 0.5;
         
-        alSourcef(source, AL_PITCH, p);
+		alSourcef(source, AL_PITCH, static_cast<float>(p));
     }
     else
     {
